requirements.txt
pandas
fpdf
yagmail
python-dotenv
openpyxl